delivery: remove panel lifecycle ( removepanelw1 )
Collapse delivery around validator evidence so sealing and merge eligibility no longer depend on panel artifacts.
